src/main/java/dev/mouhieddine/recipeapp/converters/UnitOfMeasureToUnitOfMeasureCommand.java
package dev.mouhieddine.recipeapp.converters;

import dev.mouhieddine.recipeapp.commands.UnitOfMeasureCommand;
import dev.mouhieddine.recipeapp.domain.UnitOfMeasure;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

/**
 * @author : Mouhieddine.dev
 * @since : 12/11/2020, Friday
 **/
@Component
public class UnitOfMeasureToUnitOfMeasureCommand implements Converter<UnitOfMeasure, UnitOfMeasureCommand> {
  @Synchronized
  @Nullable
  @Override
  public UnitOfMeasureCommand convert(UnitOfMeasure unitOfMeasure) {
    if (unitOfMeasure == null) return null;
    final UnitOfMeasureCommand unitOfMeasureCommand = new UnitOfMeasureCommand();
    unitOfMeasureCommand.setId(unitOfMeasure.getId());
    unitOfMeasureCommand.setDescription(unitOfMeasure.getDescription());
    return unitOfMeasureCommand;
  }
}
